feat: move Gateway card from Channels to Account settings page
